Add file existance check to file delete endpoint

+++ name_server/main.py
@@ -37,6 +37,8 @@ async def client_file_read(filename: str):
 
 @app.get("/file/delete")
 async def client_file_delete(filename: str):
+    if not check_file_existance(filename):
+        raise HTTPException(status_code=404, detail=f"File '{filename}' does not exist in that directory!")
     return file_delete(filename)
